services/notification_service/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
psycopg[binary]==3.2.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
djangorestframework-simplejwt==5.3
pytest==8.4.2
pytest-django==4.11.1
drf-gis==1.1
gunicorn==21.2